requirements.txt
# Core packages
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-learn>=1.0.0
joblib>=1.1.0
python-dotenv>=0.19.0

# API and data packages
requests>=2.26.0
ccxt>=2.0.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
pycoingecko>=2.2.0

# Trading and backtesting
backtrader>=1.9.76.123
pyfolio>=0.9.2
empyrical>=0.5.5
alpaca-trade-api>=2.1.0

# Sentiment analysis
nltk>=3.6.0
praw>=7.6.0  # Reddit API wrapper
vaderSentiment>=3.3.0

# Web server (optional, for monitoring UI)
flask>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
aiohttp>=3.8.0
gunicorn>=20.1.0

# Data processing
pytz>=2021.1

# Email notifications
secure-smtplib>=0.1.1

# Logging and monitoring
colorlog>=6.0.0

# Technical analysis
ta>=0.10.0  # Technical analysis library

# Testing
pytest>=7.0.0
pytest-mock>=3.6.0

# Development tools
jupyter>=1.0.0
jupyterlab>=3.1.0
black>=21.12b0
flake8>=4.0.0 